fix: restore extraServerBlocks as global setting
Fixes #2120

The extraServerBlocks feature was broken after refactoring to support
multiple DNS zones. It was originally added in PR #1709 as a global
setting (k8gb.coredns.extraServerBlocks) but was mistakenly moved to
per-zone in the schema without updating the template.

This commit restores the original global behavior by:
- Moving extraServerBlocks from k8gbDnsZone to root level in schema
- Updating template to access .Values.extraServerBlocks
- Removing it from per-zone configuration in values.yaml
- Adding proper documentation and example at root level
- Updating README.md to reflect the change

The global approach aligns with the original use case of adding glue
records for cluster discovery that apply to all k8gb zones.

Signed-off-by: Barrera, Angel <Angel.Barrera@adidas.com>