Document GatewayAPI integration, and include in local setup (#2119)
This PR adds an additional namespace to the local setup: `test-gslb-gatewayapi`. In this namespace we deploy podinfo and expose it using GatewayAPI HTTPRoute configuration.

Since we already had Istio installed, we use Istio as our ingress controller. This additional ingress gateway has an HTTP listener and is exposed on ports 8000 (k3d-test-gslb1), 8001 (k3d-test-gslb2).

The PR also includes a couple of fixes regarding RBAC, so that the k8gb controller can read the new gatewayAPI CRDs.

---

#### Local setup demo

Resolving DNS records:
```
➜ dig @localhost -p 1053 failover-httproute.cloud.example.com +short
172.18.0.6
172.18.0.7
➜ dig @localhost -p 1053 roundrobin-httproute.cloud.example.com +short
172.18.0.6
172.18.0.11
172.18.0.7
172.18.0.12
```

Hitting the new gateway on cluster 1:
```
➜ curl localhost:8000 -H "Host:failover-httproute.cloud.example.com"
...
  "message": "Gateway API: eu",
...
➜ curl localhost:8000 -H "Host:roundrobin-httproute.cloud.example.com"
...
  "message": "Gateway API: eu",
...
```

Hitting the new gateway on cluster 2:
```
➜ curl localhost:8001 -H "Host:failover-httproute.cloud.example.com"
...
  "message": "Gateway API: us",
...
➜ curl localhost:8001 -H "Host:roundrobin-httproute.cloud.example.com"
...
  "message": "Gateway API: us",
...
```

---

The following [documentation page](https://www.k8gb.io/resource_ref/) is extended to include GatewayAPI and Istio examples.

Finally, the example in the [frontpage](https://www.k8gb.io/) is updated to match using `name` instead of `matchLabels`, as it is more intuitive. The example still shows an Ingress resource, since GatewayAPI is not yet released.