Introduce support for IPv4 addresses in dual-stack setups (#2125)
As described in #2107, in dual stack setups the loadBalancer services expose IPv4 and IPv6 addresses.
At the moment, k8gb assumes all IP addresses are IPv4. This leads to the creation of DNSEndpoints with IPv6 addresses in the A record field:
```
apiVersion: v1
items:
- apiVersion: externaldns.k8s.io/v1alpha1
  kind: DNSEndpoint
  ...
  spec:
    endpoints:
    - dnsName: localtargets-example.gslb.aviqa.domain.com
      recordTTL: 30
      recordType: A
      targets:
      - 192.168.1.1
      - fc00:6ce6:4410::161
```

These records are then read by coreDNS, which converts the IPv6 address to `0.0.0.0`, breaking IPv4 DNS resolution.
```
dig example.gslb.domain.com +short
192.168.1.1
0.0.0.0
```

This PR does not yet introduce support for IPv6 addresses, but it makes sure IPv6 addresses are not included in the list ofexposed A records. We can therefore guarantee that k8gb works for IPv4 addresses in dual-stack setups.

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>